Added Authorization Middleware for Campground Edits and Deletions
- Created an isAuthor middleware to ensure that only the author of a campground can edit or delete it.

- Refactored code by moving the validateCampground and isAuthor middleware functions into the middleware.js file for better organization and code reuse.
